+++ Cr-8/src/main/java/com/Cr_8/dto/InfoFormRequest.java
@@ -1,4 +1,4 @@
-package com.Cr_8.controllers;
+package com.Cr_8.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
