+++ Cr-8/src/main/java/com/Cr_8/servicies/BookedDatesService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.Cr_8.dto.BookedDateDTO;
 import com.Cr_8.entities.BookedDate;
 import com.Cr_8.entities.BookingRequest;
 import com.Cr_8.entities.DayFraction;
@@ -12,8 +13,6 @@ import com.Cr_8.entities.Reference;
 import com.Cr_8.exceptions.ResourceNotFoundException;
 import com.Cr_8.repositories.BookedDatesRepo;
 import com.Cr_8.repositories.DayFractionRepo;
-
-import dto.BookedDateDTO;
 
 
 @Service
